fix: use Langfuse v4 trace propagation API
