+++ tests/llm/test_langfuse_runtime_metadata.py
@@ -1,7 +1,9 @@
 import inspect
 import logging
+from collections.abc import Mapping
+from contextlib import contextmanager
 from types import SimpleNamespace
-from typing import Any
+from typing import Any, cast
 
 import pytest
 
@@ -15,26 +17,70 @@ class FakeLangfuseClient:
     def __init__(self, *, raise_on_update: bool = False) -> None:
         self.raise_on_update = raise_on_update
         self.span_calls: list[dict[str, Any]] = []
-        self.trace_calls: list[dict[str, Any]] = []
+        self.propagate_calls: list[dict[str, Any]] = []
 
     def update_current_span(self, **kwargs: Any) -> None:
         if self.raise_on_update:
             raise RuntimeError("span boom")
         self.span_calls.append(kwargs)
 
-    def update_current_trace(self, **kwargs: Any) -> None:
+    @contextmanager
+    def propagate_attributes(self, **kwargs: Any):
         if self.raise_on_update:
             raise RuntimeError("trace boom")
-        self.trace_calls.append(kwargs)
+        self.propagate_calls.append(kwargs)
+        yield
 
 
-def test_langfuse_sdk_update_current_trace_supports_required_attrs() -> None:
+def _fake_langfuse_module(fake_client: FakeLangfuseClient) -> SimpleNamespace:
+    return SimpleNamespace(
+        get_client=lambda: fake_client,
+        propagate_attributes=fake_client.propagate_attributes,
+    )
+
+
+def test_langfuse_sdk_supports_required_span_and_trace_attrs() -> None:
+    import langfuse
     from langfuse import get_client
 
-    signature = inspect.signature(get_client().update_current_trace)
+    span_signature = inspect.signature(get_client().update_current_span)
+    for parameter in ("name", "metadata"):
+        assert parameter in span_signature.parameters
 
+    trace_signature = inspect.signature(langfuse.propagate_attributes)
     for parameter in ("user_id", "session_id", "metadata", "tags"):
-        assert parameter in signature.parameters
+        assert parameter in trace_signature.parameters
+
+
+def test_langfuse_propagate_attributes_sets_current_span_attrs() -> None:
+    from langfuse import propagate_attributes
+    from opentelemetry.sdk.trace import TracerProvider
+    from opentelemetry.sdk.trace.export import SimpleSpanProcessor
+    from opentelemetry.sdk.trace.export.in_memory_span_exporter import (
+        InMemorySpanExporter,
+    )
+
+    exporter = InMemorySpanExporter()
+    provider = TracerProvider()
+    provider.add_span_processor(SimpleSpanProcessor(exporter))
+    tracer = provider.get_tracer(__name__)
+
+    with (
+        tracer.start_as_current_span("langfuse-propagation-test"),
+        propagate_attributes(
+            user_id="user-123",
+            session_id="chat-abc",
+            tags=["honcho", "memory"],
+            metadata={"honcho_operation": "dialectic_chat"},
+        ),
+    ):
+        pass
+
+    attributes = cast(Mapping[str, object], exporter.get_finished_spans()[0].attributes)
+    assert attributes["user.id"] == "user-123"
+    assert attributes["session.id"] == "chat-abc"
+    assert attributes["langfuse.trace.tags"] == ("honcho", "memory")
+    assert attributes["langfuse.trace.metadata.honcho_operation"] == "dialectic_chat"
 
 
 def test_update_current_langfuse_observation_noops_without_public_key(
@@ -45,7 +91,7 @@ def test_update_current_langfuse_observation_noops_without_public_key(
     monkeypatch.setitem(
         __import__("sys").modules,
         "langfuse",
-        SimpleNamespace(get_client=lambda: fake_client),
+        _fake_langfuse_module(fake_client),
     )
 
     update_current_langfuse_observation(
@@ -58,7 +104,7 @@ def test_update_current_langfuse_observation_noops_without_public_key(
     )
 
     assert fake_client.span_calls == []
-    assert fake_client.trace_calls == []
+    assert fake_client.propagate_calls == []
 
 
 def test_update_current_langfuse_observation_merges_span_metadata_and_updates_trace(
@@ -70,7 +116,7 @@ def test_update_current_langfuse_observation_merges_span_metadata_and_updates_tr
     monkeypatch.setitem(
         __import__("sys").modules,
         "langfuse",
-        SimpleNamespace(get_client=lambda: fake_client),
+        _fake_langfuse_module(fake_client),
     )
 
     metadata = {
@@ -102,7 +148,7 @@ def test_update_current_langfuse_observation_merges_span_metadata_and_updates_tr
             },
         }
     ]
-    assert fake_client.trace_calls == [
+    assert fake_client.propagate_calls == [
         {
             "user_id": "user-123",
             "session_id": "chat-abc",
@@ -121,7 +167,7 @@ def test_update_current_langfuse_observation_keeps_existing_behavior_without_new
     monkeypatch.setitem(
         __import__("sys").modules,
         "langfuse",
-        SimpleNamespace(get_client=lambda: fake_client),
+        _fake_langfuse_module(fake_client),
     )
 
     update_current_langfuse_observation(
@@ -138,7 +184,7 @@ def test_update_current_langfuse_observation_keeps_existing_behavior_without_new
             }
         }
     ]
-    assert fake_client.trace_calls == []
+    assert fake_client.propagate_calls == []
 
 
 def test_update_current_langfuse_observation_is_fail_open(
@@ -150,7 +196,7 @@ def test_update_current_langfuse_observation_is_fail_open(
     monkeypatch.setitem(
         __import__("sys").modules,
         "langfuse",
-        SimpleNamespace(get_client=lambda: FakeLangfuseClient(raise_on_update=True)),
+        _fake_langfuse_module(FakeLangfuseClient(raise_on_update=True)),
     )
 
     update_current_langfuse_observation(
